+++ Example/example_variables.tf
@@ -3,39 +3,16 @@
 #
 
 #--------------------------------------------------------------
-#   Example #{MODULEDISPLAYNAME}# variables
+#   Example Service Principal variables
 #--------------------------------------------------------------
-#   / Naming
-variable "name_override" { default = null }
-variable "region_code" { default = null }
-variable "subsc_code" { default = null }
-variable "env" { default = null }
-variable "base_name" { default = null }
-variable "additional_name" { default = null }
-variable "iterator" { default = null }
-variable "owner" { default = null }
-variable "add_random" { default = null }
-variable "rnd_length" { default = null }
-variable "additional_tags" {
-  description = "(Optional) Additional base tags."
-  type        = map(string)
-  default     = null
-}
-variable "naming_values" {
-  type = object({
-    region_code     = optional(string)
-    subsc_code      = optional(string)
-    env             = optional(string)
-    base_name       = optional(string)
-    additional_name = optional(string)
-    iterator        = optional(string)
-    owner           = optional(string)
-    additional_tags = optional(map(string))
-  })
-  description = "(Optional) A terraform object with the naming values in 1 variable."
-  default     = null
-}
 
 
-#   / #{MODULEDISPLAYNAME}# specifics
-variable "#{MODULECODE}#_additional_tags" { default = null }
+
+#   / Service Principal specifics
+variable "spn_additional_tags" { default = null }
+variable "kv_id" { default = null }
+variable "test_owners" {
+  type        = list(string)
+  description = "(Optional) List of owners' object IDs. Example: `[\"00000000-0000-0000-0000-000000000000\"]`."
+  default     = []
+}