+++ data.tf
@@ -3,5 +3,6 @@
 #
 
 #--------------------------------------------------------------
-#   #{MODULEDISPLAYNAME}# module data resources
+#   Service Principal module data resources
 #--------------------------------------------------------------
+data "azurerm_client_config" "current" {}