+++ terraform_version.tf
@@ -3,7 +3,7 @@
 #
 
 #--------------------------------------------------------------
-#   #{MODULEDISPLAYNAME}# module terraform version
+#   Service Principal module terraform version
 #--------------------------------------------------------------
 terraform {
   required_version = "~> 1.0"